Final Commit push. Ready for Review.
Today i added a checkbox next to the video, i added text saying that you need to support HTML5 in order to play the video. Then i added styles to the label in order to make it a certain way. Then i took off the checkbox of the rendered page. I also added a spin to my key frames spin.